Flatten migration

+++ db/migrate/20140225150653_add_threshold_to_objective.rb
@@ -1,5 +1,5 @@
 class AddThresholdToObjective < ActiveRecord::Migration
   def change
-    add_column :objectives, :threshold, :decimal
+    add_column :objectives, :threshold, :float
   end
 end
